+++ docs-util/oas-output/operations/admin/post_admin_pricing_rule-types_[id].ts
@@ -29,8 +29,6 @@
  * tags:
  *   - Pricing
  * responses:
- *   "200":
- *     description: OK
  *   "400":
  *     $ref: "#/components/responses/400_error"
  *   "401":