+++ packages/medusa/src/api-v2/admin/stock-locations/[id]/route.ts
@@ -4,8 +4,39 @@ import {
 } from "@medusajs/utils"
 import { MedusaRequest, MedusaResponse } from "../../../../types/routing"
 
+import { deleteStockLocationsWorkflow, updateStockLocationsWorkflow } from "@medusajs/core-flows"
 import { MedusaError } from "@medusajs/utils"
-import { deleteStockLocationsWorkflow } from "@medusajs/core-flows"
+import { AdminPostStockLocationsLocationReq } from "../validators"
+
+export const POST = async (
+  req: MedusaRequest<AdminPostStockLocationsLocationReq>,
+  res: MedusaResponse
+) => {
+  const { id } = req.params
+
+  await updateStockLocationsWorkflow(req.scope).run({
+    input: {
+      selector: { id: req.params.id },
+      update: req.validatedBody,
+    },
+  })
+
+  const remoteQuery = req.scope.resolve(ContainerRegistrationKeys.REMOTE_QUERY)
+
+  const [stock_location] = await remoteQuery(
+    remoteQueryObjectFromString({
+      entryPoint: "stock_locations",
+      variables: {
+        id,
+      },
+      fields: req.remoteQueryConfig.fields,
+    })
+  )
+
+  res.status(200).json({
+    stock_location,
+  })
+}
 
 export const GET = async (req: MedusaRequest, res: MedusaResponse) => {
   const { id } = req.params