+++ packages/workflow-engine-redis/src/services/workflow-orchestrator.ts
@@ -5,11 +5,12 @@ import {
   TransactionStep,
 } from "@medusajs/orchestration"
 import { ContainerLike, Context, MedusaContainer } from "@medusajs/types"
-import { InjectSharedContext, MedusaContext, isString } from "@medusajs/utils"
+import { InjectSharedContext, isString, MedusaContext } from "@medusajs/utils"
 import {
   FlowRunOptions,
   MedusaWorkflow,
   ReturnWorkflow,
+  resolveValue,
 } from "@medusajs/workflows-sdk"
 import Redis from "ioredis"
 import { ulid } from "ulid"
@@ -75,6 +76,8 @@ export class WorkflowOrchestratorService {
   protected redisSubscriber: Redis
   private subscribers: Subscribers = new Map()
 
+  protected redisDistributedTransactionStorage_: RedisDistributedTransactionStorage
+
   constructor({
     redisDistributedTransactionStorage,
     redisPublisher,
@@ -91,11 +94,18 @@ export class WorkflowOrchestratorService {
     redisDistributedTransactionStorage.setWorkflowOrchestratorService(this)
     DistributedTransaction.setStorage(redisDistributedTransactionStorage)
 
+    this.redisDistributedTransactionStorage_ =
+      redisDistributedTransactionStorage
+
     this.redisSubscriber.on("message", async (_, message) => {
       const { instanceId, data } = JSON.parse(message)
 
       await this.notify(data, false, instanceId)
     })
+  }
+
+  async onApplicationShutdown() {
+    await this.redisDistributedTransactionStorage_.onApplicationShutdown()
   }
 
   @InjectSharedContext()
@@ -512,30 +522,39 @@ export class WorkflowOrchestratorService {
         await notify({ eventType: "onStepBegin", step })
       },
       onStepSuccess: async ({ step, transaction }) => {
-        const response = transaction.getContext().invoke[step.id]
+        const stepName = step.definition.action!
+        const response = await resolveValue(
+          transaction.getContext().invoke[stepName],
+          transaction
+        )
         customEventHandlers?.onStepSuccess?.({ step, transaction, response })
 
         await notify({ eventType: "onStepSuccess", step, response })
       },
       onStepFailure: async ({ step, transaction }) => {
-        const errors = transaction.getErrors(TransactionHandlerType.INVOKE)[
-          step.id
-        ]
+        const stepName = step.definition.action!
+        const errors = transaction
+          .getErrors(TransactionHandlerType.INVOKE)
+          .filter((err) => err.action === stepName)
+
         customEventHandlers?.onStepFailure?.({ step, transaction, errors })
 
         await notify({ eventType: "onStepFailure", step, errors })
       },
 
       onCompensateStepSuccess: async ({ step, transaction }) => {
-        const response = transaction.getContext().compensate[step.id]
+        const stepName = step.definition.action!
+        const response = transaction.getContext().compensate[stepName]
         customEventHandlers?.onStepSuccess?.({ step, transaction, response })
 
         await notify({ eventType: "onCompensateStepSuccess", step, response })
       },
       onCompensateStepFailure: async ({ step, transaction }) => {
-        const errors = transaction.getErrors(TransactionHandlerType.COMPENSATE)[
-          step.id
-        ]
+        const stepName = step.definition.action!
+        const errors = transaction
+          .getErrors(TransactionHandlerType.COMPENSATE)
+          .filter((err) => err.action === stepName)
+
         customEventHandlers?.onStepFailure?.({ step, transaction, errors })
 
         await notify({ eventType: "onCompensateStepFailure", step, errors })